Stop using swift temp url for config transport
We now download the software configs and push them to the nodes
using ansible. So there is no point of using swift temp url any more.
Though we don't use occ, it's better to set it to POLL_SERVER_HEAT
rather than the default POLL_SERVER_CFN which has a dependency on
heat-api-cfn.

Change-Id: Id7ea0c30c60e07f366e1a211a83e135b9735161f