Correct jinja loop logic for role_networks
The `_role_networks` variable in network loops was being defined and then
potentially redefined in a larger loop. If this option is first populated
with a None type and then later discovers a hash the deployment will fail
with a stack trace:

    'None' has no attribute 'append'

A full traceback from the error can be seen here[0]. To reolsve this
issue the empty array, `_role_networks` has been moved into the
`role.networks` case statement which will ensure the option is type safe
for every round.

[0] - https://pasted.tech/pastes/edba7903977c841901c0e655850d4f8b08b32e6c.raw

Closes-Bug: #1831885
Change-Id: I480786ed4d6b0a5544535ce232ae8206d04299b1
Signed-off-by: Kevin Carter <kecarter@redhat.com>