Disable ceph-dashboard by default in Stein
ceph-ansible 4 recently changed defaults and now
deploys dashboard by default. This will cause a
TripleO Stein deployment to fail unless additional
parameters are added which Stein does not support.
Opt to override this new ceph-ansible default in
Stein (master, the future Train) will support it
however.

This patch is going directly into Stein instead of
being added to master and then being backported,
since master has a CephEnableDashboard parameter
in addition to plenty of code to correctly set up
the dashboard which would be too much to backport.
Instead of misleading the user with a non-working
CephEnableDashboard parameter, we just directly
pass a flag to ceph-ansible to disable it.

Closes-Bug: #1838460

Change-Id: I718cd33513607efefb2cef3ebe110c523f6c6c6a