Point InternalTLSVncCAFile to /etc/ipa/ca.crt
In case the freeipa CA is a sub CA of an external CA the
InternalTLSVncCAFile requrested does not have the full CA
chain and only have the free IPA CA. As a result qemu
which can not verify the vnc certificate sent by the
vnc-proxy. The issue is in certmonger[1] as it does not return the
full CA chain.

As a workaround, until certmonger is fixed, this change points the
InternalTLSVncCAFile to /etc/ipa/ca.crt which has the full CA chain.

[1] - https://bugzilla.redhat.com/show_bug.cgi?id=1710632

Change-Id: Id2392dd076e216b4cd5b405d893ec2959456673e
(cherry picked from commit ade09f3a3405f47fa9524ef4baea155a4262175b)