Fix missing syslog name in libvirt log setting
The commit I6e5ac7f5971df5998db22c5d392084fddf77d34a  makes libvirt
directly send logs to rsyslog to avoid duplicated header, but in fact
it was not working properly because of missing syslog name specified.

This patch fixes the wrong value passed there, so that we have
the log surely directed to syslog.

Change-Id: I246d0f365c1b8eef7730a1abe0507555e23e6408