Update mysql url for cell0 in step3
When doing a TLS brown field deployment (ie. updating a previously
deployed system without internal TLS to using internal TLS), the
database endpoint stored in the db for the default cell and cell0
must be changed in step3.

Logic already exists in step 3 to change the default cell.  Logic
has been added to change cell0.

Note that this is not a problem in stein+ because we have already
added logic to set the db endpoint for both cells using
templated values for both cells in step 3.

Closes-Bug: 1838804
Change-Id: I3a0bc2ba62c9c0a2f4860a5b362f59251e858fcb