Update mysql url for cell0 in step3
When doing a TLS brown field deployment (ie. updating a previously
deployed system without internal TLS to using internal TLS), the
database endpoint stored in the db for the default cell and cell0
must be changed in step3.

Logic already exists in step 3 to change the default cell, and the
FFU group added logic to update cell0 in step 5.  We just need to
move that logic to step 3.

Note that this is not a problem in stein+ because we have already
added logic to set the db endpoint for both cells using
templated values for both cells in step 3.

Closes-Bug: 1838804

Change-Id: I7b5f6e0a2c8ba77fd575cf1a1003a1553f96efff