Redis HA TLS: do not use the pacemaker image tag for redis_tls_proxy
When running a 3-nodes Redis cluster with TLS, we use an intermediate
container redis_tls_proxy for all the encrypted communication between
the redis nodes.

The intermediate container is managed by paunch, so it shouldn't use
the container image tagged :pcmklatest. This tag is a intermediate
tag only meant to be consumed by pacemaker to implement container
update without service disruption. When used in paunch, it causes
the deployment to fail at step 2.

Configure redis_tls_proxy to use heat parameter ContainerRedisImage
instead. That is the image meant to be consumed by paunch containers.

Closes-Bug: 1839305

Change-Id: Id4bba3df475582fa3884e38d17ca66913da83657