Force "Pre-cache" tasks to run in dry run
We don't want the pre-cache tasks to be skipped in dry-run so we force
check_mode to "no".

For example, "command" is skipped in check mode because we don't provide
"creates" nor "removes"; and since we need to cache facts when running Puppet
in dry mode, we need to make sure this task is run anyway.

See https://docs.ansible.com/ansible/latest/modules/command_module.html

For "file", "copy", and "synchronize", they aren't skipped in dry mode
but the resources aren't executed for real.

So this patch disable check_mode for all these tasks.
We also move the tasks in a block which avoid tags & check_mode
duplication for each task and avoid problems in the future if a task is
added.

Closes-Bug: #1839321
Change-Id: I21524419b8b416d0879fc61aa3274d2e30537438