Add tags always into external update tasks.
After bumping the Ansible version in Tripleo to 2.6, it was needed
to perform a change in the include module into include_tasks or
import_tasks, as include was getting deprecated [0]. The external
update/upgrades tasks got impacted by that change, but as they use
a loop to execute the tasks we couldn't use import_tasks.
The way include_tasks handles the tasks execution depending on the
tags differs from import_tasks (dynamic vs static) and as a consequence
when running the external upgrade run passing --tags container_image_prepare
we didn't see any tasks running. This behavior got fix in [1], which
seems to be the right way to preserve the import_tasks tags handling as
explained in [2]. But the external update tasks were missed to patch.

This patch includes the tags: always statement inside the external
update tasks and also syncs the variables content we pass into external_update_tasks
as many of those variables, which we do pass for the external_upgrade_playbook,
were missing in the external_update_playbook.

[0] - https://review.opendev.org/#/c/579844/
[1] - https://review.opendev.org/#/c/639642/1
[2] - https://github.com/ansible/ansible/issues/30882#issuecomment-380596557
Closes-Bug: #1839520
Change-Id: If7e7b4bbb3cead0887384cc543ce37e9ee5396ab
(cherry picked from commit 6c675af9babf475cbc825b437b37ff06de4b461b)