Add the ability to configure ovn-remote-probe-interval
This parameter sets inactive probe interval of the JSON
session from ovn-controller to the OVN SB database.
By default this it is 5s which not be sufficient in
loaded systems or during high control-plane activity spikes,
leading to unnecessary reconnections to OVSDB server.
Now it is extended by default to 1 min and it
is configurable.

Depends-On: https://review.opendev.org/#/c/675918
Change-Id: Ie7cb761ad3b4a180990de2916d6210d15ec0bf50
Closes-Bug: #1836604
(cherry picked from commit 942dd1367eb1e379528569e719e5b9df45684857)