Force re-run of pacemaker bundle init containers during upgrade-scaleup
We need to recreate the *_init_bundle containers during the staged
upgrade (upgrade-scaleup), because the init containers will e.g. set
the correct replica count for the bundle as we're scaling up from 1 to
3 controllers.

The init containers wouldn't get re-run/recreated automatically on
upgrade-scaleup because none of their input would change. The
DeploymentIdentifier is set by the `openstack upgrade prepare` command
and applied during upgrade with `--limit controller-0`. Then later
when we run `--limit controller-0,controller-1`, the
DeploymentIdentifier value does not change from the perspective of
controller-0, so we have to explicitly remove the bundle init
containers to force re-running them.

Change-Id: I93c6d7292f37529735e24286f4076634aa4fe6a5
Partial-Bug: #1831690