Fix NovaResumeGuestsStateOnHostBoot when using podman
There is an issue if NovaResumeGuestsStateOnHostBoot gets enabled on the
initial deploy run. It tries to start libvirt-guests in the
host_prep_tasks task. While this is not an issue with docker where we run
the libvirt-guests direct on the host (outside a container), with podman
we run the libvirt-guests inside the container and in the initial deploy
we don't have the container at this stage.

This disables to start the tripleo_nova_libvirt_guests service in the
host_prep_tasks task in case of podman is used.

Change-Id: I7569ea69399807ce2efcc6ca9403db166181a2db
Closes-Bug: #1840618