Explicitly set notification driver for novajoin
Novajoin is reliant on some notifications from nova. We recently merged
an update to the undercloud installation that disables the notification
drvier across the board if telemetry is disabled. This change adds a
service_config_setting for nova to force the notification driver to
messagingv2. This should be merged with the service configurations and
override the value that is normally set via NotificationDriver.

Change-Id: Ifee8f7f50fc2ef99605cbbcaa89016514b80b561
Closes-Bug: #1841100
(cherry picked from commit 7c9c363c789811d54353d769aab86774eb5c96f8)