Ensure container_startup_configs_json_stat is defined
When using --tags to run only a subset of tasks from
common/deploy-steps-tasks.yaml, the condition that checks the result of
container_startup_configs_json_stat.stat.exists was failing since the
task that defined it was skipped.

This patch adds an additional "is defined" check to ensure the var is
defined be the result is checked.

Change-Id: Iadde90ed9416902848df2e60551470c0f1689a32