Sort bootstrap node check for container_puppet_tasks
Previously, the task that checks if container_puppet_tasks should be
written to the bootstrap node was not sorted, which meant the order was
undefined coming out of the Ansible inventory.

This patch adds a sort filter to the check so that by convention the
first node is chosen from the inventory.

A variable, sorted_bootstrap_node, is also added to control the sorting
in the bootstrap_node.j2 template from the tripleo-hieradata role so
that the tripleo-ansible patch is able to merge first.

Depends-On: I3d595ea5b84f940a3b2dbc69798f69fe03529c10

Change-Id: I6b93f5b0747c5a11d24615a3bbb5516f9be81401