Remove HostPrepConfig (for config-download)
HostPrepConfig was using the old way (Heat) to run Ansible. We don't
need it anymore when using config-download.

This is an equivalent backport of
I2041e6892de76b0ed04d7497e3f9064bfaf58270 (why the same change-id is
used here). That patch can not just be backported as-is to stable/queens
because the non-config-download mechanism still needs to work, so we can
not just remove the resources.

Instead, we provide a new mapping that is set to OS::Heat::None when
config-download is activated with the environment file.

Change-Id: I2041e6892de76b0ed04d7497e3f9064bfaf58270