Fix Manila when run without pacemaker
The supported path is to run the manila-share service
under control of pacemaker using container bundles started
in Step5.  However, it is useful for test and POC purposes
to be able to run the manila-share service without pacemaker
control as a regular container started in Step4.

This change fixes the non-pacemaker deployment path, which runs
manila-share in Step4 rather than Step5, so that it uses the same
privileges for the manila-share container as are used for the
pacemaker-controlled container-bundle in Step5.

However. we do not change scenario004 to deploy manila using the
non-pacemaker path since in CI we want to test the path that is
supported for customers.

Change-Id: I8effea96a4428281b5a5dfa9711bb602fdbdc705
Related-Bug: #1841941
(cherry picked from commit cff553c7e651aef769ef0ef7f0fc5e9757467cce)
(cherry picked from commit 68a4e125bbd02e7ae5f695045c64201e4a7b5c83)