container-puppet: run podman rm with --storage
There is an ongoing bug in libpod where the removal of a container can
randomly (probably under heavy load) leak and fail to finish the tasks,
leaving leftovers in the podman database, which causes a container
creation with the same name to fail.

It is related to this issue:
https://github.com/containers/libpod/issues/3906

And originally reported here as well:
https://bugzilla.redhat.com/show_bug.cgi?id=1747885

This patch is a mitigation that has been validated by the Container Team
where we use --storage option to force the low level removal of the
container by using containers/storage project to really remove all the
data.

Closes-Bug: #1840691
Change-Id: I711f460bd51747c3985b95784d19560d1be2028a
(cherry picked from commit 9e2a9715b5379221bd027b77a4989d87f7c9d88f)