SELinux: correct type for /var/log/containers
The correct type for this directory is "container_file_t" (or
svirt_sandbox_file_t). The var_log_t was needed before in order to allow
syslog to write HAProxy logs in /var/log/containers/haproxy.

This is not needed anymore, since a patch in openstack-selinux[1] allows
syslog to have a full access to container_file_t type.
Moreover, since we have logrotate running in a container, it mounts the
/var/log/containers location with ":z" flag, which re-labels all the
files to container_file_t.

[1] https://github.com/redhat-openstack/openstack-selinux/commit/f9b45cede356c0e09f78ffb8a3c01fb80b6aac36

Change-Id: I13a90695686b9134f6fcceac1bf6d22c2ac390a5