Run facts gathering always for upgrades.
Up to now, each of the upgrade steps to gather facts from the
Undercloud and Overcloud nodes did contain a tag named facts.
This allowed us to run the facts gathering at will, however
there is a problem with the upgrade_tasks. As it is allowed to
pass a specific tag during its execution. So, if the task passed
is different than facts, for example system_upgrade, then the
facts gathering tasks wont be executed and the upgrade tasks
will fail when trying to access some of the ansible facts (as
it could be ansible_hostname).
The solution for this is to run always the facts gathering, no
matter what tag is being passed we'll have the facts available
during the execution.

Change-Id: Ibee072bec9916804163dab29e164eb1423cd08f8