Correct SELinux type for host openvswitch logs
Openvswitch server is running on the host, and puts its logs in
/var/log/openvswitch, while its agents are in containers and put logs in
/var/log/containers/openvswitch.

This means the /var/log/openvswitch container doesn't need to get the
fancy "container_file_t" type, and can be set to the right one,
openvswitch_log_t.

This will prevent issues with different softwares, such as logrotate or
even openvswitch itself on a SELinux enforcing system.

Change-Id: I4a786ecb60190759754d17f7b4e84d93f7ffb389