Add "--rm" to the container "run" command
Doing so should help preventing the "ghost container" effect we
described in #1840691.
It is still possible to keep failed containers, if we pass
"keep_container: True" to the container-puppet.py script env.

We still keep the "container_rm()" method, since it apparently can help
ensuring all is actually removed in the end.

Change-Id: I1029fb24491b867616293f8487abc2b5295bca57
Related-Bug: #1840691