Move common step 1 tasks to their own file
The large block of tasks in deploy-steps-tasks.yaml that are run only at
step 1, results in a large number of tasks being skipped at steps 2-5.
Task skipping is not "free", as ansible has to repeatedly compute what
needs to be skipped.

Note that the previous usage with the block statement causes the "when"
statement to be inherited to each task, so it has to be recalculated for
every task across every node for all steps.

The issue starts to become apparent with deployments with larger number
of nodes.

This commit refactors the steps 1 tasks into their own tasks file, and
then uses the include_task module to only include the tasks at step 1.

A separate play is also used for these step 1 tasks so that ansible
doesn't have to repeatedly calculate if the include_task should be
executed for steps 2-5 as well.

Change-Id: Id5985ce8ac741baa9adc9f5874df0459fd4c24b2