Resolve broken zaqar container caused by logging issues
The zaqar container is broken due to the log file being owned by root. When
the zaqar-server log file is unabnle to be written to by the zaqar process
it causes a traceback resulting in 500 errors. This change ensures that the
zaqar log directory has the proper permissions and that the log file within
the directory is created when the container is started. A sticky bit is
being used on the zaqar log directory to ensure all files created within
the directory retain group expected permissions in almost all circumstances.

Related-Bug: #1843875

Change-Id: I63442f0bdec11179c361f503906166f75c5e0355
Signed-off-by: Kevin Carter <kecarter@redhat.com>
(cherry picked from commit 4579d42e8d4e0f651f6c12137fb2c61d002b37ad)