Remove containers before removing associated storage
We added --storage option in https://review.opendev.org/#/c/679793/
when removing containers with podman. However, it won't delete the exited
containers, i.e only orphaned storage would be removed with it.

So, we have to do both 'rm' and 'rm --storage' with podman to ensure
that we don't get the container name already used error,

Change-Id: I66ce37b19dd855e9798a7877494264f6a128d284
Depends-On: https://review.opendev.org/681989
Related-Bug: #1840691