Move Octavia agent containers to step 5
Octavia uses external deploy steps to complete configuration of the
support services, requiring a restart to pick these changes up if the
services are started in step 4. This patch moves the startup of these
services to step 5 avoiding the need for restarting.

This was actually causing an issue with healthchecks as the restart was
happening during the restart.

Change-Id: I4d7d322c2d64ed06b71ab0da049cf92f5a8e8d8a
Related-Bug: #1843981