[Queens] Run cellv2 host discovery via workflow or deploy_steps_tasks
This is a queens only modified version of
https://review.opendev.org/682644 due to the different deployment
methods we have there.

When the compute bootstrap node gets blacklisted or deployment is
run with skip-deploy-identifier the discovery job we have in step_5
won't run. This runs the the discovery via an ansible playbook to
make sure it runs in this conditions.

In case of default deploy method the discovery is run by the
discovery workflow.

In case of config-download the discovery is run via the
deploy_steps_tasks.

Change-Id: Icda4c1eb3e8c39a01547586dbc6f0407ce846c64
Closes-Bug: #1831711
Depends-On: I54d42df162a6744806301d97bca5d94e5f380a2b