Create an endpoint entry for novajoin
On a nova-less undercloud, the novajoin endpoint needs to be
discoverable by a mistral workflow without any configuration file
entry.

This change does this by creating an endpoint entry for novajoin.
There is already a keystone service entry created for novajoin, so
this change creates the endpoint which probably should have existed
anyway.  novajoin is accessed directly instead of through haproxy, so
the endpoint URL is created manually instead of through EndpointMap.

Change-Id: I0d307bec3771c7b03cbad2bb8ac0c1f20eb9baf3
Blueprint: nova-less-deploy