Move nested bash script into a real file
container-puppet.py had within it a nested bash script using
a string literal. This change moves that script into a stand
alone file which will be copied to the host and used when
needed. This change will begin allowing us to deconstruct
the container-puppet.py process allowing us to simplify the
process, inturn making deployments more reliable, faster,
and easier to understand.

Change-Id: I447f33f159c5a50e55401b4e0a893c250869d185
Signed-off-by: Kevin Carter <kecarter@redhat.com>