Override policy for undercloud novajoin
The default policy compute_service_user[1] limits novajoin calls to
the nova user. This prevents novajoin being called from a mistral
workflow, which is required for a nova-less undercloud.

This change overrides this policy so that any user with the admin role
can call novajoin. Service users including the nova user have the
admin role, so this should not affect the existing nova based novajoin
calls.

[1] https://opendev.org/x/novajoin/src/branch/master/novajoin/policy.py#L39
Depends-On: https://review.opendev.org/#/c/684435/
Change-Id: I9034c1bce9c296ff183f0d7f2e22df187699fa04
Blueprint: nova-less-deploy