Add DHCPv6 rules + set ipversion for dhcp rules
Use the ipversion parameter for firewall rules to contain
rule creation in either iptables or ip6tables. Add rules
in ironic-inspector and neutron deployment template to
add rules for DHCPv6 in ip6tables.

DHCPv6 relay and DHCPv6 server both use port 547 so 547
need to be open for both INPUT and OUTPUT.

Related-bug: #1845153
Depends-On: Id872c55cfc6b958fef3ccda2d923f821a1fe6a13
Depends-On: I8b453f7c13c2015aa208ed1bddcdca246cdca58d
Change-Id: If91b883459488856ae54e3ca0d0fb97d4d248f97