Resolve FIXME for NeutronServer mapping
The NeutronServer service mapping has been outdated for a while in favor
of using the NeutronApi service name. Previously, the 2 service mappings
just pointed at the same service template.

This patch properly cleans up any old files still using NeutronServer,
and adds a deprecation for the old mapping. An environment file is added
to enable the old mapping if needed.

This patch will be backported to train, and then the environment file
can be removed in ussuri in a subsequent patch.

Also cleans up an outdated FIXME comment in
overcloud-resource-registry.j2.yaml.

Change-Id: I48b969941c3c6cb1a5b39c8292f461c61b3dabd7