Use single quotes for --start-at-task debug tasks
Previously, the use of double quotes in these tasks caused ansbile to
escape the double quote when printing the debug message. The escape
character (\) made copy/pasting the task name more difficult.

Switching to single quotes around the task name makes it easier on the
operator to copy/paste so that --start-at-task can be easily used.

Change-Id: I5e5837cc22462769de05be73249633de4fa65fcf