Add NovaAZConfig service to compute roles
The NovaAZConfig service is added to all the compute related roles
of the box so that the nodes are automatically in a new AZ when the
service is enabled.

By default, the service is mapped to OS::Heat::None so no change in
default behavior is made.

Change-Id: If1e4e149100dccfe76ccd807997a611d0fc8841f
(cherry picked from commit 45c226c8fe9e0da6f62841608a2c635d7eec8ad8)