Undercloud hosts entries in overcloud nodes
With IPv6 the overcloud nodes cannot pull images from the
undercloud using an IPv6 address. By adding hosts entries
for the undercloud a name can be used instead of the IPv6
address.

The parameter 'UndercloudHostsEntries' is populated with
entries from the HEAT_HOSTS section of /etc/hosts on the
undercloud.

Related-Bug: #1836578
Depends-On: Id9271460b20e602f935f1eddd70f989cd605b1f4
Change-Id: Iac6efde9dd283906274d95c3a239b4b882ec052e