Add posibilities to configure OVNNorthboundServerPort in split stacks
This changes is need by split stacks when ovn dbs are not colocated
with neutron-server to configure correctly OVNNorthboundServerPort.
When neutron api is deployed with ovn dbs on the same node,
ovn dbs template set this port. But in other cases when
they are not colocated we need to set this port in neutron
plugin template.

Change-Id: Ic84cc8864c799eaa4273c022b07445ec997a7fc9
Signed-off-by: Kamil Sambor <ksambor@redhat.com>