Remove all old pcmklatest images
An operator may make a mistake when importing images, resulting
in more than one image being present. This may happen, for example,
if an image is imported from one source, then another.

To ensure that all old images are deleted, and to prevent a failure
in the command executing the deletion if there is more than one
result when listing the images, we process the deletion in a loop
per stdout line.

Note:
This patch is not required in more recent branches due to other
changes in the process which have eliminated the possibility of
there being more than one image to delete.

Closes: rhbz#1727173
Change-Id: I67e0a321fe2c77e54966cd9b70be56a489dd2007