Fix placement_wait_for_service
This fix the indent and volumes of the placement_wait_for_service
and the corresponding placement_wait_for_service.py to use the
config of the extracted placement service.

It also
* changes to set placement::keystone::authtoken::auth_url
instead of placement::keystone::authtoken::auth_uri as auth_uri is
deprecated and not supported by placement::keystone::authtoken.
* sets placement::keystone::authtoken::region_name

Related-Bug: 1842948

Change-Id: Ic24cf646efdd70ba1dbca42d3408847fe09a6e49