Generate startup configs files per step and per container
For each step and each container, generate a unique JSON, and Paunch
will be able to read them all thanks to the patch in dependency
(tripleo-ansible).

Old location:
/var/lib/tripleo-config/container-startup-config-step1.json
We keep the old files for backward compatibility.

New location:
/var/lib/tripleo-config/container-startup-config/step_1/haproxy.json

Note: hashed files won't be generated for the old location anymore,
since it's done via container-puppet.py in which we now give the new
location.

Story: 2006732
Task: 37162
Depends-On: If0f1c6c308cd58f7baa9a8449fbf685ff10f0e0a
Change-Id: I1cf8923a698d0f6e0b1e00a7985f363a83e914c4