Fix update of non HA container during update.
After the refactoring of step_1 to improve speed during deployment, we
were not updating the /var/lib/container-puppet/container-puppet.json
or the container-startup-configs.json for instance.

Net result is that the configuration is being done on the initially
deployed container and non HA container are not updated.

Note that not all step1 tasks were refarctored into the
deploy_steps_tasks_step_1. So we still need to loop over step 1 in the
deploy_steps_tasks.

Closes-Bug: #1848359

Change-Id: I76b16cd7781ea601778004afa4e0bc3020ce3c59