Add configurable monitor timeouts for ovn dbs
Under pressure, the default monitor timeout value of 20 seconds is not
enough to prevent unnecessary failovers of the ovn-dbs pacemaker resource.
While spawning a few VMs in the same time this could lead to unnecessary
movements of master DB, then re-connections of ovn-controllers (slaves are
read-only), further peaks of load on DBs, and at the end it could lead to
snowball effect. Now this value can be configurable by
OVNDBSPacemakerTimeout which will configure
tripleo::profile::pacemaker::ovn_dbs_bundle (default is set to 60s)

Depends-On: https://review.opendev.org/#/c/689166/
Signed-off-by: Kamil Sambor <ksambor@redhat.com>

Change-Id: I1afb5f2ef31ec61b3b224e5e1672fb9f12bcb110