memcached: do not run memcached from a bash process
memcached is run from a bash script that sources various
environment variables from config file. If the container
doesn't use dumb-init, bash is run as pid 1, but in
such case it is unable to catch SIGTERM.

Make the bash script exec into memcached rather than just
spawning memcached as a child. This way, memcached becomes
the pid 1 when dumb-init is not in used. This ensures
that a SIGTERM can be caught when issuing docker/podman
stop.

Change-Id: I15677fcf5ffc3913c1818bae691cc7548be0af0e
Closes-Bug: #1848710